requirements.txt
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
stripe==12.4.0b2
python-dotenv==1.1.1
google-cloud-secret-manager==2.23.3
pandas==2.3.1
google-cloud-logging==3.12.1
colorama==0.4.6
singleton-decorator==1.0.0
redis==6.4.0